+++ legacy/client/jest.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-  testEnvironment: 'node',
-  transform: {
-    '^.+\\.ts?$': 'ts-jest',
-  }
-};