+++ sol-did/programs/sol-did/src/lib.rs
@@ -2,7 +2,7 @@
 //!
 //! It exposes the public #program instructions, as well as integrations libraries that are
 //! independent of the anchor framework.
-
+#![allow(clippy::result_large_err)]
 #![warn(
 unused_import_braces,
 unused_imports,
@@ -12,12 +12,12 @@ unused_imports,
 )]
 
 pub mod constants;
+pub mod errors;
 pub mod integrations;
 pub mod legacy;
 pub mod state;
 pub mod utils;
 
-mod errors;
 mod instructions;
 mod security_txt;