+++ sol-did/programs/sol-did/src/security_txt.rs
@@ -1,6 +1,7 @@
 pub mod security {
     use solana_security_txt::security_txt;
 
+    #[cfg(not(feature = "no-entrypoint"))]
     security_txt! {
         // Required fields
         name: "Sol-DID",